refactor(lib, api): remove last references to request in lib, move to api
Removes the `RequestUnmarshaller` interface, which was only being used for `GetParams` in a sugar endpoint. Using the interface means having to define "special" parsing directions in `lib`, rather than keeping them in `api` where they belong, closer to the sugar handlers that user them.

Moved all the methods that parse from the request into one file.
