+++ lib/datasets.go
@@ -8,7 +8,6 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
-	"net/http"
 	"path/filepath"
 	"regexp"
 	"strings"
@@ -22,7 +21,6 @@ import (
 	"github.com/qri-io/jsonschema"
 	"github.com/qri-io/qfs"
 	"github.com/qri-io/qfs/localfs"
-	"github.com/qri-io/qri/api/util"
 	"github.com/qri-io/qri/automation/run"
 	"github.com/qri-io/qri/base"
 	"github.com/qri-io/qri/base/archive"
@@ -121,49 +119,6 @@ func isValidSelector(selector string) bool {
 }
 
 var validSelector = regexp.MustCompile(`^$|^[\w-\.]*[\w]$`)
-
-// UnmarshalFromRequest satisfies the Unmarshaller interface
-func (p *GetParams) UnmarshalFromRequest(r *http.Request) error {
-	log.Debugf("GetParams.UnmarshalFromRequest ref:%s", r.FormValue("ref"))
-	p.Ref = r.FormValue("ref")
-
-	ref, err := dsref.Parse(p.Ref)
-	if err != nil {
-		return err
-	}
-
-	if ref.Username == "me" {
-		return fmt.Errorf("username \"me\" not allowed")
-	}
-
-	p.Selector = r.FormValue("selector")
-
-	p.All = util.ReqParamBool(r, "all", true)
-	p.Limit = util.ReqParamInt(r, "limit", 0)
-	p.Offset = util.ReqParamInt(r, "offset", 0)
-	if !(p.Offset == 0 && p.Limit == 0) {
-		p.All = false
-	}
-
-	if p.All && p.Limit == 0 && p.Offset == 0 {
-		page := -1
-		pageSize := -1
-		if i := util.ReqParamInt(r, "page", 0); i != 0 {
-			page = i
-		}
-		if i := util.ReqParamInt(r, "pageSize", 0); i != 0 {
-			pageSize = i
-		}
-		// we don't want the defaults to override this and also only want to
-		// set values if they are present
-		if page >= 0 && pageSize >= 0 {
-			p.Limit = pageSize
-			p.Offset = (page - 1) * pageSize
-			p.All = false
-		}
-	}
-	return nil
-}
 
 // GetResult returns the dataset or some part of it as structured data inside the `Value` field
 // The `Bytes` field is reserved for data that can only be expressed as a slice of bytes
