+++ lib/datasets_test.go
@@ -431,115 +431,6 @@ func TestGetParamsSetNonZeroDefaults(t *testing.T) {
 	}
 }
 
-func TestGetParamsUnmarshalFromRequest(t *testing.T) {
-	cases := []struct {
-		description  string
-		url          string
-		expectParams *GetParams
-		muxVars      map[string]string
-	}{
-		{
-			"get request with ref",
-			"/get/peer/my_ds",
-			&GetParams{
-				Ref: "peer/my_ds",
-				All: true,
-			},
-			map[string]string{"ref": "peer/my_ds"},
-		},
-		{
-			"get request with a selector",
-			"/get/peer/my_ds/meta",
-			&GetParams{
-				Ref:      "peer/my_ds",
-				Selector: "meta",
-				All:      true,
-			},
-			map[string]string{"ref": "peer/my_ds", "selector": "meta"},
-		},
-		{
-			"get request with limit and offset",
-			"/get/peer/my_ds/body",
-			&GetParams{
-				Ref:      "peer/my_ds",
-				Selector: "body",
-				Limit:    0,
-				Offset:   10,
-			},
-			map[string]string{"ref": "peer/my_ds", "selector": "body", "limit": "0", "offset": "10"},
-		},
-		{
-			"get request with 'all'",
-			"/get/peer/my_ds/body",
-			&GetParams{
-				Ref:      "peer/my_ds",
-				Selector: "body",
-				All:      true,
-			},
-			map[string]string{"ref": "peer/my_ds", "selector": "body", "all": "true"},
-		},
-	}
-	for _, c := range cases {
-		t.Run(c.description, func(t *testing.T) {
-			r := httptest.NewRequest("GET", c.url, nil)
-			mustSetMuxVarsOnRequest(t, r, c.muxVars)
-			gotParams := &GetParams{}
-			err := gotParams.UnmarshalFromRequest(r)
-			if err != nil {
-				t.Error(err)
-				return
-			}
-			if diff := cmp.Diff(c.expectParams, gotParams); diff != "" {
-				t.Errorf("output mismatch (-want +got):\n%s", diff)
-			}
-		})
-	}
-
-	badCases := []struct {
-		description string
-		url         string
-		expectErr   string
-		muxVars     map[string]string
-	}{
-		{
-			"get me",
-			"/get/me/my_ds",
-			`username "me" not allowed`,
-			map[string]string{"ref": "me/my_ds"},
-		},
-		{
-			"bad parse",
-			"/get/peer/my+ds",
-			`unexpected character at position 7: '+'`,
-			map[string]string{"ref": "peer/my+ds"},
-		},
-	}
-	for i, c := range badCases {
-		t.Run(c.description, func(t *testing.T) {
-			r := httptest.NewRequest("GET", c.url, nil)
-			mustSetMuxVarsOnRequest(t, r, c.muxVars)
-			gotParams := &GetParams{}
-			err := gotParams.UnmarshalFromRequest(r)
-			if err == nil {
-				t.Errorf("case %d: expected error, but did not get one", i)
-				return
-			}
-			if diff := cmp.Diff(c.expectErr, err.Error()); diff != "" {
-				t.Errorf("output mismatch (-want +got):\n%s", diff)
-			}
-		})
-	}
-}
-
-func mustSetMuxVarsOnRequest(t *testing.T, r *http.Request, muxVars map[string]string) {
-	q := r.URL.Query()
-	for varName, val := range muxVars {
-		q.Add(varName, val)
-	}
-	r.URL.RawQuery = q.Encode()
-	return
-}
-
 func TestGetZip(t *testing.T) {
 	ctx, done := context.WithCancel(context.Background())
 	defer done()
